Feature/1202138782029385 update consumer balance on account screen

+++ lib/common/models/account.dart
@@ -20,7 +20,7 @@ class CEVAccount {
   String? address;
   String? did;
   String? seed;
-  BigInt? balance;
+  double? balance;
   BigInt? tokenDecimals;
   String? tokenSymbol;
 }
